+++ packages/elements/src/utils/is-base64/index.ts
@@ -8,12 +8,3 @@ export const isBase64 = (base64Str: string | undefined | null) => {
     return false
   }
 }
-
-export const getTypeFromBase64 = (base64: string): string => {
-  try {
-    const type = base64.split(';')[0].split(':')[1]
-    return type
-  } catch (error) {
-    return ''
-  }
-}
