+++ packages/elements/src/utils/validators/validate-text-and-number.ts
@@ -3,8 +3,3 @@ export const isTextAndNumberOnly = (value?: string): boolean => {
   const re = /^[a-zA-Z0-9|\s]+$/
   return re.test(value)
 }
-
-export const isImageType = (type: string) => {
-  const regex = /^image\//
-  return regex.test(type)
-}
