+++ packages/utils/index.ts
@@ -1,3 +1,5 @@
 export * from './src/catch-chunk-error'
 export * from './src/sentry-logger'
 export * from './src/object'
+export * from './src/base64'
+export * from './src/validators'
