+++ packages/utils/src/base64.ts
@@ -1,0 +1,8 @@
+export const getTypeFromBase64 = (base64: string): string => {
+  try {
+    const type = base64.split(';')[0].split(':')[1]
+    return type
+  } catch (error) {
+    return ''
+  }
+}
