+++ packages/utils/src/validators/__tests__/string.test.ts
@@ -1,0 +1,10 @@
+import { isImageType } from '../string'
+
+describe('isImageType', () => {
+  it('should return true', () => {
+    const input = 'image/jpeg'
+    const output = true
+    const result = isImageType(input)
+    expect(result).toEqual(output)
+  })
+})
