+++ packages/utils/src/validators/string.ts
@@ -1,0 +1,4 @@
+export const isImageType = (type: string) => {
+  const regex = /^image\//
+  return regex.test(type)
+}
